Rent/src/test/java/com/example/Rent/Controllers/ReservationControllerTest.java
package com.example.Rent.Controllers;

import com.example.Rent.Models.Reservation;
import com.example.Rent.Repository.ReservObjectRepository;
import com.example.Rent.Repository.ReservationRepository;
import com.example.Rent.Repository.TenantRepository;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import java.util.Date;
import java.util.LinkedList;
import java.util.List;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;


@WebMvcTest(ReservationController.class)
class ReservationControllerTest {

    @Autowired
    private MockMvc mvc;


    @MockBean
    ReservationRepository reservationRepository;

    @MockBean
    ReservObjectRepository reservObjectRepository;

    @MockBean
    TenantRepository tenantRepository;

    @Test
    void GetByTent() throws Exception{
        String uri = "/api/Reservation/getByTent/ColbXF";
        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();

        int status = mvcResult.getResponse().getStatus();
        assertEquals(200, status);
    }
}